StaticMPI

Documentation for StaticMPI.

StaticMPI.MPI_Comm_rank

StaticMPI.MPI_Comm_size

StaticMPI.MPI_Finalize

StaticMPI.MPI_Init

StaticMPI.MPI_Comm_rank — Method

MPI_Comm_rank(comm)

Obtain the rank of the calling MPI task within the specified MPI communicator comm

StaticMPI.MPI_Comm_size — Method

MPI_Comm_size(comm)

Obtain the number of tasks in the specified MPI communicator comm.

StaticMPI.MPI_Finalize — Method

MPI_Finalize()

Conclude the execution of the calling MPI task.

StaticMPI.MPI_Init — Method

MPI_Init()
MPI_Init(argc::Int, argv::Ptr{Ptr{UInt8}})

Initialize the execution environment of the calling MPI task for single- threaded execution. The optional arguments argc and argv are the traditional argument count and argument value pointer variables passed to a compiled executable by the operating system.

Returns MPI_SUCCESS on success